merge

+++ src/main/java/net/proselyte/springsecurityapp/dao/StepDao.java
@@ -7,5 +7,5 @@ import java.util.List;
 
 public interface StepDao extends JpaRepository<Step,Long> {
     Step findById(int id);
-    List<Step> findAllByInstructionsId(int id);
+    List<Step> findAllByInstructionsId(Long id);
 }
